pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gguf-clone"
version = "0.1.0"
description = "Create optimized quants by using similar GGUFs as templates"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "spicyneuron" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "numpy~=1.26.4",
  "sentencepiece~=0.2.0",
  "transformers>=4.57.1,<5.0.0",
  "protobuf>=4.21.0,<5.0.0",
  "gguf @ git+https://github.com/ggml-org/llama.cpp.git#subdirectory=gguf-py",
  "huggingface-hub>=0.24.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0.1",
  "torch~=2.6.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "pytest>=8.4.2",
    "ruff>=0.14.11",
]

[project.scripts]
gguf_clone = "gguf_clone.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
